// video-aggregator/scripts/types.ts
export interface Thumbnail {
  url: string;
  width?: number;
  height?: number;
}

export interface VideoThumbnails {
  default?: Thumbnail;
  medium?: Thumbnail;
  high?: Thumbnail;
  standard?: Thumbnail;
  maxres?: Thumbnail;
}
// ...
export interface Video {
  title: string;
  publishedAt: string;
  thumbnails: VideoThumbnails;
  videoUrl: string;
  duration: string;
  durationSeconds?: number;
  privacyStatus: string;
  description: string;
}